Rename <crispy/logger.h> to <crispy/debuglog.h>
