+++ src/crispy/CMakeLists.txt
@@ -27,7 +27,7 @@ target_sources(crispy-core INTERFACE
     ${CMAKE_CURRENT_SOURCE_DIR}/compose.h
     ${CMAKE_CURRENT_SOURCE_DIR}/escape.h
     ${CMAKE_CURRENT_SOURCE_DIR}/indexed.h
-    ${CMAKE_CURRENT_SOURCE_DIR}/logger.h
+    ${CMAKE_CURRENT_SOURCE_DIR}/debuglog.h
     ${CMAKE_CURRENT_SOURCE_DIR}/overloaded.h
     ${CMAKE_CURRENT_SOURCE_DIR}/reference.h
     ${CMAKE_CURRENT_SOURCE_DIR}/span.h
